+++ include/init.php
@@ -7,3 +7,4 @@ declare(strict_types=1);
 
 require_once(__DIR__ . '/config.php');
 require_once(__DIR__ . '/config.defaults.php');
+require_once(__DIR__ . '/i18n.php');